Fixed validators variant & added dataChecker

+++ programming1/5lab/src/data/classes/Validators/classes/CoordinateValidator/CoordinateXYValidator.java
@@ -2,9 +2,9 @@ package src.data.classes.Validators.classes.CoordinateValidator;
 
 import src.data.classes.Validators.interfaces.ValidatorInterface;
 
-public class CoordinateXYValidator implements ValidatorInterface<Double> {
+public class CoordinateXYValidator implements ValidatorInterface<Long> {
     @Override
-    public boolean validate(Double variable) {
+    public boolean validate(Long variable) {
         return (variable != null && variable > -212);
     }
 }
