+++ programming1/5lab/src/data/classes/Validators/classes/CoordinateValidator/CoordinatesValidator.java
@@ -5,8 +5,8 @@ import src.data.classes.Validators.interfaces.ValidatorInterface;
 public class CoordinatesValidator implements ValidatorInterface<String[]> {
     @Override
     public boolean validate(String[] args) {
-        if (new CoordinateXYValidator().validate(Double.parseDouble(args[0])) &&
-        new CoordinateXYValidator().validate(Double.parseDouble(args[1]))) {
+        if (new CoordinateXYValidator().validate(Long.parseLong(args[0])) &&
+        new CoordinateXYValidator().validate(Long.parseLong(args[1]))) {
             return true;
         }
         return false;
