+++ programming1/5lab/src/data/classes/Validators/classes/HumanValidator/HeightHumanValidator.java
@@ -2,9 +2,9 @@ package src.data.classes.Validators.classes.HumanValidator;
 
 import src.data.classes.Validators.interfaces.ValidatorInterface;
 
-public class HeightHumanValidator implements ValidatorInterface<Integer> {
+public class HeightHumanValidator implements ValidatorInterface<Float> {
     @Override
-    public boolean validate(Integer height) {
+    public boolean validate(Float height) {
         return height > 0;
     }
 }
