+++ programming1/5lab/src/data/classes/Validators/classes/HumanValidator/HumanValidator.java
@@ -6,7 +6,7 @@ public class HumanValidator implements ValidatorInterface<String[]>{
     @Override
     public boolean validate(String[] args) {
         if (new AgeHumanValidator().validate(Integer.parseInt(args[0])) &&
-        new HeightHumanValidator().validate(Integer.parseInt(args[1]))) {
+        new HeightHumanValidator().validate(Float.parseFloat(args[1]))) {
             return true;
         }
         return false;
